==> notebooks/1_1_stage1_feature_engg.ipynb <==
# some parts have been taken from https://www.kaggle.com/the1owl/redefining-treatment-0-57456

import pandas as pd
import numpy as np
import feather
from sklearn import preprocessing as pe
from tqdm import tqdm

from utils import text_to_wordlist, get_fp

def get_wt(cls):
    tot_neg_instances = df_train.shape[0] - p[cls]
    tot_pos_instances = p[cls]
    return float(tot_neg_instances/tot_pos_instances)

D = 2 ** 24
def do_feature_engg(df_all):
    
    # create a new col GeneVar
    df_all['GeneVar'] = df_all['Gene'] + ' ' + df_all['Variation']
    
    # get the share of the gene expression in the corresponding text in the row
    df_all['Gene_Share'] = df_all.apply(lambda row: sum([1 for w in row['Gene'].split(' ') 
                                                       if w in row['Text'].split(' ')]), axis=1)


    # get the share of the variation in the corresponding text in the row
    df_all['Variation_Share'] = df_all.apply(lambda row: sum([1 for w in row['Variation'].split(' ') 
                                                            if w in row['Text'].split(' ')]), axis=1)

    # get the share of the gene-variation in the corresponding text in the row
    df_all['Gene_Variation_Share'] = df_all.apply(lambda row: sum([1 for w in row['GeneVar'].split(' ') 
                                                            if w in row['Text'].split(' ')]), axis=1)

    # get the gene length
    df_all['GL'] = df_all['Gene'].apply(lambda x: len(x))

    # get the variation length
    df_all['VL'] = df_all['Variation'].apply(lambda x: len(x))

    # print(sorted(df_all['VL'].values))

    max_len=55 # VL has max length 55 
    for i in range(max_len+1):
        df_all['Gene_'+str(i)] = df_all['Gene'].map(lambda x: str(x[i]) if len(x)>i else '')
        df_all['Variation_'+str(i)] = df_all['Variation'].map(lambda x: str(x[i]) if len(x)>i else '')
        df_all['GeneVar_'+str(i)] = df_all['GeneVar'].map(lambda x: str(x[i]) if len(x)>i else '')

    # from https://www.kaggle.com/the1owl/redefining-treatment-0-57456
#     gene_var_lst = sorted(list(df_train.Gene.unique()) + list(df_train.Variation.unique()))
    gene_var_lst = sorted(list(df_all.Gene.unique()) + list(df_all.Variation.unique()))
    gene_var_lst = [x for x in gene_var_lst if len(x.split(' '))==1]

    i_ = 0

    for el in tqdm(gene_var_lst):
        df_all['GV_'+str(el)] = df_all['Text'].map(lambda x: str(x).count(str(el)))
        i_ += 1
    
    # gene expression value counts
    s = df_all['Gene'].value_counts()
    df_all['G_VC'] = df_all['Gene'].apply(lambda gene: s[str(gene)])
    # print(df_all.G_VC[df_all.Gene == 'BRCA1'].count())

    # variation value counts
    s = df_all['Variation'].value_counts()
    df_all['V_VC'] = df_all['Variation'].apply(lambda var: s[str(var)])

    # gene variation value counts
    s = df_all['GeneVar'].value_counts()
    df_all['GV_VC'] = df_all['GeneVar'].apply(lambda var: s[str(var)])

    # hash gene variation and text
    df_all['hash'] = df_all['Gene'] + '_' + df_all['Variation'] + '_' + df_all['Text']
    df_all['hash'] = df_all['hash'].map(lambda x: hash(x) % D)

    # label encoding, length of column and number of words in the column
    for c in tqdm(df_all.columns):
        if df_all[c].dtype == 'object':
            if c in ['Gene','Variation','GeneVar']:
                lbl = pe.LabelEncoder()
                df_all[c+'_lbl_enc'] = lbl.fit_transform(df_all[c].values)  
                df_all[c+'_len'] = df_all[c].map(lambda x: len(str(x)))
                df_all[c+'_words'] = df_all[c].map(lambda x: len(str(x).split(' ')))
            elif c != 'Text':
                lbl = pe.LabelEncoder()
                df_all[c] = lbl.fit_transform(df_all[c].values)
            if c=='Text': 
                df_all[c+'_len'] = df_all[c].map(lambda x: len(str(x)))
                df_all[c+'_words'] = df_all[c].map(lambda x: len(str(x).split(' '))) 

    return df_all


df_train = feather.read_dataframe('../cache/train_stage1.feather')

df_train.shape

df_train.head()

df_test = feather.read_dataframe('../cache/test_stage1.feather')

df_test.shape

df_test.head()

y = df_train['Class'].values
df_train = df_train.drop(['Class'], axis=1)

test_ids = df_test['ID'].values

df_all = pd.concat((df_train, df_test), axis=0, ignore_index=True)

df_all.head()

df_all.shape

D = 2 ** 24

df_all = do_feature_engg(df_all)

df_train = df_all.iloc[:len(df_train)]
df_test = df_all.iloc[len(df_train):]

df_train['Class'] = y

df_train.shape

df_test.shape

df_test.head()

feather.write_dataframe(df_train, '../cache/train_stage1_fe.feather')

df_train.head()

df_train['hash'].value_counts()

feather.write_dataframe(df_test, '../cache/test_stage1_fe.feather')

df_train['Text'] = [text_to_wordlist(w) for w in df_train['Text'].values]
df_test['Text'] = [text_to_wordlist(w) for w in df_test['Text'].values]

test_id = df_test.ID.values
df = pd.DataFrame()
df['ID'] = test_id
df.to_csv('../cache/stage1_test_id.csv', index=False)

y = df_train.Class.values
df = pd.DataFrame()
df['y'] = y
df.to_csv('../cache/stage1_labels.csv', index=False)

p =pd.value_counts(df_train['Class'])

df_train['wt'] =  df_train['Class'].map(lambda s: get_wt(s))

wt = df_train.wt.values
df = pd.DataFrame()
df['wt'] = wt
df.to_csv('../cache/stage1_weights.csv', index=False)
np.save('../cache/stage1_train_weights', wt)

my_wt = {}
n_class = 9
for cls in range(n_class):
    my_wt[cls+1] = get_wt(cls+1)
np.save('../cache/stage1_train_weights_per_class', my_wt)

fp = 0
fp = get_fp(fp)

df_train = df_train.drop('Class',axis=1)
df_train = df_train.drop('wt',axis=1)

df_train = fp.fit_transform(df_train)
print (df_train.shape)

df_test = fp.fit_transform(df_test)
print (df_test.shape)

np.save('../cache/train_stage1_fe2', df_train)
np.save('../cache/test_stage1_fe2', df_test)

df_train = feather.read_dataframe('../cache/train_stage1_fe.feather')
df_test = feather.read_dataframe('../cache/test_stage1_fe.feather')

df_train.head()

print(df_train.shape, df_test.shape)

df_stage1_partial_soln = pd.read_csv('../data/stage1_solution_filtered.csv')
cols = []
for cls in range(9):
    cols.append('class' + str(cls+1))
df_stage1_partial_soln['Class'] = df_stage1_partial_soln[cols].idxmax(axis=1).map(lambda x: int(x[-1]))

partial_id = df_stage1_partial_soln['ID'].values
df_partial_test = df_test[df_test.ID.isin(partial_id)].reset_index()

df_stage1_partial_soln.head()

df_all_p = pd.concat((df_train, df_partial_test), axis=0, ignore_index=True)

df_all_p = do_feature_engg(df_all_p)

df_train2 = df_all_p.iloc[:len(df_train2)]
df_test2 = df_all_p.iloc[len(df_train2):]

feather.write_dataframe(df_train2, '../cache/train_p_stage1_fe.feather')
feather.write_dataframe(df_test2, '../cache/test_p_stage1_fe.feather')

df_train2 = feather.read_dataframe('../cache/train_p_stage1_fe.feather')
df_test2 = feather.read_dataframe('../cache/test_p_stage1_fe.feather')

print(df_train2.shape, df_test2.shape)

set(df_train2.columns) - set(df_test2.columns)

set(df_test2.columns) - set(df_train2.columns)

df_train2 = df_train2.drop('Class',axis=1)
df_test2 = df_test2.drop('Class',axis=1)

df_train2 = df_train2.drop('index',axis=1)
df_test2 = df_test2.drop('index',axis=1)

df_train2.head()

df_train2['Text'] = [text_to_wordlist(w) for w in df_train2['Text'].values]
df_test2['Text'] = [text_to_wordlist(w) for w in df_test2['Text'].values]

test_id2 = df_test2.ID.values
df2 = pd.DataFrame()
df2['ID'] = test_id2
df2.to_csv('../cache/stage1_p_test_id.csv', index=False)



fp = 0
fp = get_fp(fp)

df_train2 = fp.fit_transform(df_train2)
print (df_train2.shape)

df_test2 = fp.fit_transform(df_test2)
print (df_test2.shape)

tr = df_train2
# te = df_test2

print(np.isnan(df_train2).any())
print(np.isnan(df_train2).any())

print(np.argwhere(np.isnan(df_train2)))

df_train2.shape

df_test2.shape

# df_train2[:,~np.all(np.isnan(df_train2), axis=0)].shape

# df_test2[:,~np.all(np.isnan(df_test2), axis=0)].shape

np.save('../cache/train_p_stage1_fe2', df_train2)
np.save('../cache/test_p_stage1_fe2', df_test2)

df_stage1_partial_soln.head()

y1 = df_stage1_partial_soln.Class.values
df2 = pd.DataFrame()
df2['y'] = y1
df2.to_csv('../cache/stage1_p_test_labels.csv', index=False)
