==> notebooks/2_0_stage1_xgb.ipynb <==
import re
import numpy as np
import pandas as pd
import feather
import xgboost as xgb
import feather

from sklearn.model_selection import train_test_split
from sklearn.metrics import log_loss, confusion_matrix, f1_score, accuracy_score

from utils import print_confusion_matrix

df_train = np.load('../cache/train_stage1_fe2.npy')

df1 = pd.read_csv('../cache/stage1_labels.csv')
y = df1['y'].values

df_test = np.load('../cache/test_stage1_fe2.npy')
df = pd.read_csv('../cache/stage1_test_id.csv')
pid = df.ID.values

df_test.shape

df_train.shape

y = y - 1 #fix for zero bound array

wts_per_class = np.load('../cache/stage1_train_weights_per_class.npy')
wts_per_class = wts_per_class.tolist()

print(wts_per_class)

 
-

submission = pd.DataFrame(preds/denom, columns=['class'+str(c+1) for c in range(9)])
submission['ID'] = pid
# submission.to_csv('../submissions/sub_stage1_all_2_2_xgb.csv', index=False)
submission.to_csv('../submissions/sub_stage1_all_2_2_wt_xgb.csv', index=False)

# score 0.4571 on stage1 public leader board

df_train = np.load('../cache/train_p_stage1_fe2.npy')

df1 = pd.read_csv('../cache/stage1_labels.csv')
y = df1['y'].values

df_test = np.load('../cache/test_p_stage1_fe2.npy')
df = pd.read_csv('../cache/stage1_p_test_id.csv')
pid = df.ID.values

print(df_train.shape, df_test.shape)


df = pd.read_csv('../cache/stage1_p_test_labels.csv')
test_labels = df['y'].values-1

test_labels.shape

y = y - 1 #fix for zero bound array

denom = 0
fold = 10 
for i in range(fold):
    params = {
        'eta': 0.03333,
        'max_depth': 6,
        'subsample' : 0.8,
        'colsample_bytree':0.8,
        'objective': 'multi:softprob',
        'eval_metric': 'mlogloss',
        'num_class': 9,
        'seed': i,
        'silent': True
    }
    x1, x2, y1, y2 = train_test_split(df_train, y, test_size=0.2, random_state=i)
    w1 = [wts_per_class[j+1] for j in y1]
    w2 = [wts_per_class[j+1] for j in y2]
    watchlist = [(xgb.DMatrix(x1, y1,weight=w1), 'train'), (xgb.DMatrix(x2, y2,weight=w2), 'valid')]
    model = xgb.train(params, xgb.DMatrix(x1, y1,weight=w1), 1000,  watchlist, verbose_eval=50, early_stopping_rounds=100)
    score1 = log_loss(y2, model.predict(xgb.DMatrix(x2), 
                                        ntree_limit=model.best_ntree_limit),  
                      labels = list(range(9)))
    print('fold = {:d}'.format(i))
    print('val multi_log_loss: {}'.format(score1))
    
    #if score < 0.9:
    if denom != 0:
        pred = model.predict(xgb.DMatrix(df_test), ntree_limit=model.best_ntree_limit+80)
        preds += pred
    else:
        pred = model.predict(xgb.DMatrix(df_test), ntree_limit=model.best_ntree_limit+80)
        preds = pred.copy()
    
    denom += 1
    score2 = log_loss(test_labels, pred, labels = list(range(9)))
    print('test multi_log_loss: {}'.format(score2))
    
    fscore = f1_score(test_labels, pred.argmax(axis=1), labels = list(range(9)), average='macro')
    print('test f1_score: {}'.format(fscore))
    
    acc = accuracy_score(test_labels, pred.argmax(axis=1))
    print('test accuracy: {}'.format(acc))
    
    print(confusion_matrix(test_labels, pred.argmax(axis=1), labels = list(range(9))))
    
    print('-------------------')
    print('\n\n')
    
#     submission = pd.DataFrame(pred, columns=['class'+str(c+1) for c in range(9)])
#     submission['ID'] = pid
#     submission.to_csv('../submissions/sub_stage1_2_2_wt_xgb_fold_'  + str(i) + '.csv', index=False)

score2 = log_loss(test_labels, pred, labels = list(range(9)))
print('test multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, pred.argmax(axis=1), labels = list(range(9)), average='macro')
print('test f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, pred.argmax(axis=1))
print('test accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, pred.argmax(axis=1), labels = list(range(9))))

# with no lambda, no alpha we get
# val multi_log_loss: 1.0247328862547875
# test multi_log_loss: 1.41869956478679
# test f1_score: 0.2549920698056178
# test accuracy: 0.4429347826086957

# with alpha: 100 we get:
# val multi_log_loss: 1.1108881736468328
# test multi_log_loss: 1.5692192874848843
# test f1_score: 0.27029986646550364
# test accuracy: 0.40217391304347827

# with alpha 500 we get:
# test multi_log_loss: 1.9443259673274083
# test f1_score: 0.2713471712458167
# test accuracy: 0.30978260869565216
    
# below with lambda 700
# test multi_log_loss: 1.45

pred1 = preds/denom

score2 = log_loss(test_labels, pred1, labels = list(range(9)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, pred1.argmax(axis=1), labels = list(range(9)), average='macro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, pred1.argmax(axis=1))
print('final accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, pred1.argmax(axis=1), labels = list(range(9))))
