==> notebooks/2_1_stage1_RF.ipynb <==
import re
import numpy as np
import pandas as pd
import feather
import xgboost as xgb
import feather

from sklearn.model_selection import train_test_split, StratifiedKFold
from sklearn.metrics import log_loss

from sklearn.cross_validation import *
from sklearn.grid_search import GridSearchCV

from sklearn.ensemble import RandomForestClassifier
from sklearn.multioutput import MultiOutputClassifier

from sklearn.utils import shuffle
from sklearn.metrics import log_loss, confusion_matrix, f1_score, accuracy_score
import pickle

df_train = np.load('../cache/train_stage1_fe2.npy')
df_test = np.load('../cache/test_stage1_fe2.npy')
df = pd.read_csv('../cache/stage1_labels.csv')
y = df['y'].values

df1 = pd.read_csv('../cache/stage1_test_id.csv')
test_id = df1['ID'].values

np.unique(y)

wts_per_class = np.load('../cache/stage1_train_weights_per_class.npy')
wts_per_class = wts_per_class.tolist()
print(wts_per_class)

param_grid = { 
    'n_estimators': [50, 200, 700],
    'max_features': ['auto', 'sqrt', 'log2']
}

model_RF = RandomForestClassifier(random_state=0, class_weight=wts_per_class)
clf0 = GridSearchCV(model_RF, param_grid=param_grid, scoring='neg_log_loss',cv=5,n_jobs=4, refit=True)
clf0.fit(df_train, y)
fname = '../cache/stage1_rf_wt_gridcv_ll_clf.pkl'
save_classifier(fname, clf0)

print(clf0.best_score_)
print(clf0.best_params_)

y

param_grid = { 
    'n_estimators': [50, 200, 700],
    'max_features': ['auto', 'sqrt', 'log2']
}

model_RF = RandomForestClassifier(random_state=0, class_weight=wts_per_class)
clf = GridSearchCV(model_RF, param_grid=param_grid, scoring='f1_macro',
                   cv=StratifiedKFold(y),n_jobs=-1, refit=True)
clf.fit(df_train, y)
fname = '../cache/stage1_rf_wt_gridcv_f1_clf.pkl'
save_classifier(fname, clf)

print(clf.best_score_)
print(clf.best_params_)

param_grid = { 
    'n_estimators': [50, 200, 700],
    'max_features': ['auto', 'sqrt', 'log2']
}

model_RF = RandomForestClassifier(random_state=0, class_weight=wts_per_class)
clf = GridSearchCV(model_RF, param_grid=param_grid, scoring='accuracy',cv=StratifiedKFold(y),n_jobs=-1)
clf.fit(df_train, y)
fname = '../cache/stage1_rf_wt_gridcv_acc_clf.pkl'
save_classifier(fname, clf)

print(clf.best_score_)
print(clf.best_params_)

df_train = np.load('../cache/train_p_stage1_fe2.npy')
df1 = pd.read_csv('../cache/stage1_labels.csv')
y = df1['y'].values

df_test = np.load('../cache/test_p_stage1_fe2.npy')
df = pd.read_csv('../cache/stage1_p_test_id.csv')
pid = df.ID.values

df = pd.read_csv('../cache/stage1_p_test_labels.csv')
test_labels = df['y'].values

df_train.shape

df_test.shape

len(y)

param_grid = { 
    'n_estimators': [700, 7000, 12000],
    'max_features': ['auto', 'sqrt', 'log2']
}

model_RF = RandomForestClassifier(random_state=0, class_weight=wts_per_class)
clf = GridSearchCV(model_RF, param_grid=param_grid, scoring='neg_log_loss',
                   cv=StratifiedKFold(y, 5, shuffle=True),n_jobs=4, 
                   refit=True)
clf.fit(df_train, y)
fname = '../cache/stage1_rf_wt_gridcv_ll_p_clf.pkl'
save_classifier(fname, clf)

print(clf.best_score_)
print(clf.best_params_)

preds_proba = clf.predict_proba(df_test)

score2 = log_loss(test_labels, preds_proba, labels = list(range(1,10)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, preds_proba.argmax(axis=1)+1, labels = list(range(1,10)), average='macro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, preds_proba.argmax(axis=1)+1)
print('final accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, preds_proba.argmax(axis=1)+1, labels = list(range(1,10))))


param_grid = { 
    'n_estimators': [700, 1000, 1200],
    'max_features': ['auto', 'sqrt', 'log2']
}

model_RF = RandomForestClassifier(random_state=0, class_weight=wts_per_class)
clf = GridSearchCV(model_RF, param_grid=param_grid, scoring='f1_weighted',cv=StratifiedKFold(y, 5),n_jobs=4, refit=True)
clf.fit(df_train, y)
fname = '../cache/stage1_rf_wt_gridcv_f1m_p_clf.pkl'
save_classifier(fname, clf)

print(clf.best_score_)
print(clf.best_params_)

preds_proba = clf.predict_proba(df_test)

score2 = log_loss(test_labels, preds_proba, labels = list(range(1,10)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, preds_proba.argmax(axis=1)+1, labels = list(range(1,10)), average='macro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, preds_proba.argmax(axis=1)+1)
print('final accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, preds_proba.argmax(axis=1)+1, labels = list(range(1,10))))


param_grid = { 
    'n_estimators': [700, 7000, 12000],
    'max_features': ['auto', 'sqrt', 'log2']
}

model_RF = RandomForestClassifier(random_state=0, class_weight=wts_per_class)
clf = GridSearchCV(model_RF, param_grid=param_grid, scoring='accuracy',cv=StratifiedKFold(y, 5),n_jobs=4, refit=True)
clf.fit(df_train, y)
fname = '../cache/stage1_rf_wt_gridcv_acc_p_clf.pkl'
save_classifier(fname, clf)

print(clf.best_score_)
print(clf.best_params_)

preds_proba = clf.predict_proba(df_test)

score2 = log_loss(test_labels, preds_proba, labels = list(range(1,10)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, preds_proba.argmax(axis=1)+1, labels = list(range(1,10)), average='macro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, preds_proba.argmax(axis=1)+1)
print('final accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, preds_proba.argmax(axis=1)+1, labels = list(range(1,10))))


param_grid = { 
    'n_estimators': [700, 7000, 12000],
    'max_features': ['auto', 'sqrt', 'log2']
}

model_RF = RandomForestClassifier(random_state=0, class_weight=wts_per_class)
clf = GridSearchCV(model_RF, param_grid=param_grid, scoring='recall_weighted', 
                   cv=StratifiedKFold(y, 5),n_jobs=4, refit=True)
clf.fit(df_train, y)
fname = '../cache/stage1_rf_wt_gridcv_cm_p_clf.pkl'
save_classifier(fname, clf)

print(clf.best_score_)
print(clf.best_params_)

preds_proba = clf.predict_proba(df_test)

score2 = log_loss(test_labels, preds_proba, labels = list(range(1,10)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, preds_proba.argmax(axis=1)+1, labels = list(range(1,10)), average='macro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, preds_proba.argmax(axis=1)+1)
print('final accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, preds_proba.argmax(axis=1)+1, labels = list(range(1,10))))

preds_proba[0].shape

save_classifier('../cache/stage1_multi_rf.pkl', multi_target_forest)
