==> notebooks/2_2_stage1_multiclassifers.ipynb <==
import re
import numpy as np
import pandas as pd
import feather
import xgboost as xgb

from sklearn.multiclass import OneVsRestClassifier
from sklearn.multiclass import OneVsOneClassifier
from sklearn.multiclass import OutputCodeClassifier
  
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
  
from sklearn.svm import LinearSVC
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import Perceptron
from sklearn.tree import DecisionTreeClassifier
from sklearn.grid_search import GridSearchCV
from sklearn import svm
from sklearn.linear_model import Perceptron
from sklearn.naive_bayes import MultinomialNB

from sklearn.utils.testing import assert_array_equal
from sklearn.utils.testing import assert_equal
from sklearn.utils.testing import assert_almost_equal
from sklearn.utils.testing import assert_true
from sklearn.utils.testing import assert_false
from sklearn.utils.testing import assert_raises
  
from sklearn.utils.testing import assert_greater
import pickle

from sklearn.metrics import log_loss, confusion_matrix, f1_score, accuracy_score
from sklearn.calibration import CalibratedClassifierCV

from utils import multi_log_loss


def save_classifier(fname, clf):
    # save the classifier
    with open(fname, 'wb') as fid:
        pickle.dump(clf, fid)

def load_classifier(fname):
    # load it again
    with open(fname, 'rb') as fid:
        clf = pickle.load(fid)
        return clf

df_train = np.load('../cache/train_p_stage1_fe2.npy')

df_test = np.load('../cache/test_p_stage1_fe2.npy')

df = pd.read_csv('../cache/stage1_labels.csv')

df.head()

y = df['y'].values

df_train.shape

df1 = pd.read_csv('../cache/stage1_p_test_labels.csv')
test_labels = df1['y'].values

wts_per_class = np.load('../cache/stage1_train_weights_per_class.npy')
wts_per_class = wts_per_class.tolist()
print(wts_per_class)

pd.value_counts(y)

2 * (953/3321) * 0.11/ ((953/3321) + 0.11)

pd.value_counts(test_labels)

df_test.shape

2 * (101/368) * 0.11/ ((101/368) + 0.11)

# def test_ovo_gridsearch():
ovo = OneVsOneClassifier(svm.SVC(probability=True))
Cs = [0.01, 0.02, 0.05, 0.09]
cv = GridSearchCV(ovo, {'estimator__C': Cs})
cv.fit(df_train, y)
best_C = cv.best_estimator_.estimators_[0].C
assert_true(best_C in Cs)


print(best_C)

ovo = CalibratedClassifierCV(OneVsOneClassifier(svm.SVC(probability=True, C=0.01)))

ovo.fit(df_train, y)

test_p = ovo.predict_proba(df_test)

test_preds = test_p

score2 = log_loss(test_labels, test_preds, labels = range(1,10))
print('stage1 partial multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, test_preds.argmax(axis=1)+1, labels = list(range(1,10)), average='micro')
print('stage1 partial f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, test_preds.argmax(axis=1)+1)
print('stage1 partial accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, test_preds.argmax(axis=1)+1, labels = list(range(1,10))))

np.unique(df['y'].values)

clf = OneVsRestClassifier(CalibratedClassifierCV(LinearSVC())).fit(df_train, df['y'].values)

test_preds = clf.predict_proba(df_test)

print(test_preds)

score2 = log_loss(test_labels, test_preds, labels=[1,2,3,4,5,6,7,8,9])
print('stage1 partial multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, test_preds.argmax(axis=1)+1, labels = list(range(1,10)), average='micro')
print('stage1 partial f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, test_preds.argmax(axis=1)+1)
print('stage1 partial accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, test_preds.argmax(axis=1)+1, labels = list(range(1,10))))

Y = df['y'].values

ovo = MultinomialNB()
Y_pred =ovo.fit(abs(df_train), Y).predict_proba(df_test)

Y_pred = Y_pred.clip(min=0.05, max=0.95)

score2 = log_loss(test_labels, Y_pred, labels=range(1,10))
print('stage1 partial multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, Y_pred.argmax(axis=1)+1, labels = list(range(1,10)), average='micro')
print('stage1 partial f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, Y_pred.argmax(axis=1)+1)
print('stage1 partial accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, Y_pred.argmax(axis=1)+1, labels = list(range(1,10))))
