==> notebooks/2_3_stage1_mlp.ipynb <==
import numpy as np
import pandas as pd

from sklearn.metrics import log_loss, confusion_matrix, f1_score, accuracy_score
from sklearn.model_selection import GridSearchCV, StratifiedKFold

from sklearn.linear_model import SGDClassifier
from sklearn.neural_network import MLPClassifier

df_train = np.load('../cache/train_p_stage1_fe2.npy')
df_test = np.load('../cache/test_p_stage1_fe2.npy')
df = pd.read_csv('../cache/stage1_labels.csv')
y = df['y'].values

df1 = pd.read_csv('../cache/stage1_p_test_labels.csv')
test_labels = df1['y'].values

wts_per_class = np.load('../cache/stage1_train_weights_per_class.npy')
wts_per_class = wts_per_class.tolist()
print(wts_per_class)

y = y-1

wts_per_class2 = {}
wts_per_class2 = {i:wts_per_class[i+1] for i in y}

# SGDClassifier(alpha=0.0001, average=False, class_weight=None, epsilon=0.1,
#        eta0=0.0, fit_intercept=True, l1_ratio=0.15,
#        learning_rate='optimal', loss='hinge', max_iter=None, n_iter=None,
#        n_jobs=1, penalty='l2', power_t=0.5, random_state=None,
#        shuffle=True, tol=None, verbose=0, warm_start=False)
# sgd_model = SGDClassifier(class_weight=wts_per_class2)

sgd_model = SGDClassifier()

# parameters = {
#     'alpha': [0.0001, 0.0005, 0.001, 0.01, 0.1], 
#     'epsilon':[0.1, 0.5, 0.05],
#     'eta0': [0.0, 0.0001, 0.001],
#     'l1_ratio': [0.01, 0.1, 0.15, 0.2],
#     'loss': ['log'],
#     'random_state': [1337]
# }

parameters = {
    'alpha': [0.001, 0.01], 
    'max_iter': [50,250,500],
    'loss': ['log'],
    'penalty': ['l2', 'l1', 'elasticnet'],
    'power_t': [0.05, 0.1],
    'random_state': [1337],
}

p = {
    'alpha': [0.001], 
    'loss': ['log'], 
    'max_iter': [50], 
    'penalty': ['elasticnet'], 
    'power_t': [0.05], 
    'random_state': [1337]}



clf0 = GridSearchCV(sgd_model, p, n_jobs=1, 
                   cv=2,
                   scoring = 'neg_log_loss',
                   verbose=2, refit=True)



clf0.fit(df_train, y)

print(clf0.best_score_)
print(clf0.best_params_)
# -14.8125892222
# {'alpha': 0.001, 'loss': 'log', 'max_iter': 250, 'random_state': 1337}

# parameters = {
#     'alpha': [0.001, 0.0001], 
#     'max_iter': [50,250,500],
#     'loss': ['log'],
#     'penalty': ['l2', 'l1', 'elasticnet'],
#     'power_t': [0.1, 0.3, 0.5],
#     'random_state': [1337],
# }
# -14.4717413019
# {'alpha': 0.001, 'loss': 'log', 'max_iter': 50, 'penalty': 'elasticnet', 'power_t': 0.1, 'random_state': 1337}
# In [14]:

# -14.4717413019
# {'alpha': 0.001, 'loss': 'log', 'max_iter': 50, 'penalty': 'elasticnet', 'power_t': 0.05, 'random_state': 1337}

test_preds = clf0.predict_proba(df_test)
print(test_preds)

score2 = log_loss(test_labels, test_preds+1, labels = list(range(1,10)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, test_preds.argmax(axis=1)+1, labels = list(range(1,10)), average='micro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, test_preds.argmax(axis=1)+1)
print('final accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, test_preds.argmax(axis=1)+1, labels = list(range(1,10))))

# earlier:
# parameters = {
#     'alpha': [0.001, 0.0001, 0.00001], 
#     'max_iter': [50,250,500,1000,5000],
#     'loss': ['log'],
#     'penalty': ['L2', 'L1', 'elasticnet'],
#     'power': [0.1, 0.03, 0.5]
#     'random_state': [1337],
# }
# final multi_log_loss: 21.11335970669253
# final f1_score: 0.2554347826086957
# final accuracy: 0.2554347826086957
# [[ 94   0   0   0   0   0   0   0   0]
#  [ 46   0   0   0   0   0   0   0   0]
#  [  7   0   0   0   0   0   0   0   0]
#  [ 65   0   0   0   0   0   0   0   0]
#  [ 25   0   0   0   0   0   0   0   0]
#  [ 22   0   0   0   0   0   0   0   0]
#  [101   0   0   0   0   0   0   0   0]
#  [  2   0   0   0   0   0   0   0   0]
#  [  6   0   0   0   0   0   0   0   0]]
# In [ ]:


mlp_model = MLPClassifier()

# MLPClassifier(hidden_layer_sizes=(100, ), activation=’relu’, solver=’adam’, alpha=0.0001, 
#               batch_size=’auto’, learning_rate=’constant’, learning_rate_init=0.001, 
#               power_t=0.5, max_iter=200, shuffle=True, random_state=None, tol=0.0001, 
#               verbose=False, warm_start=False, momentum=0.9, nesterovs_momentum=True,
#               early_stopping=False, validation_fraction=0.1,
#               beta_1=0.9, beta_2=0.999, epsilon=1e-08)
parameters = {
    'solver': ['adam', 'sgd'], 
    'max_iter': [50,250,500],
    'learning_rate_init': [0.001, 0.0001, 0.01, 0.1]
    'alpha': [0.001],
    'power_t': [0.1],
    'random_state': [1337],
}

clf = GridSearchCV(mlp_model, parameters, n_jobs=4, 
                   cv=5,
                   scoring = 'neg_log_loss',
                   verbose=2, refit=True)



clf.fit(df_train, y)

print(clf.best_score_)
print(clf.best_params_)

# -2.19642899615
# {'alpha': 0.001, 'max_iter': 50, 'power_t': 0.1, 'random_state': 1337, 'solver': 'sgd'}

test_preds = clf.predict_proba(df_test)
print(test_preds)

print(test_preds.argmax(axis=1)+1)

score2 = log_loss(test_labels, test_preds, labels = list(range(1,10)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(test_labels, test_preds.argmax(axis=1)+1, labels = list(range(1,10)), average='macro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(test_labels, test_preds.argmax(axis=1)+1)
print('final accuracy: {}'.format(acc))

print(confusion_matrix(test_labels, test_preds.argmax(axis=1)+1, labels = list(range(1,10))))
