==> notebooks/4_0_stage2_prepare_data.ipynb <==
import pandas as pd
import numpy as np
import feather
from imblearn.over_sampling import SMOTE
from sklearn import preprocessing as pe
from tqdm import tqdm

train_stage1 = pd.read_csv('../data/training_variants')
test_stage1 = pd.read_csv('../data/test_variants')
trainx_stage1 = pd.read_csv('../data/training_text', sep="\|\|", engine='python', header=None, skiprows=1, names=["ID","Text"])
testx_stage1 = pd.read_csv('../data/test_text', sep="\|\|", engine='python', header=None, skiprows=1, names=["ID","Text"])

test_stage1_soln = pd.read_csv('../data/stage1_solution_filtered.csv')

test_stage1_soln.head()

stage1_soln_ids = test_stage1_soln['ID']

test_stage1_soln = test_stage1_soln.drop('ID',axis=1)

test_stage1_soln.shape

test_stage1_soln['class'] = test_stage1_soln.idxmax(axis=1)

test_stage1_soln['class'] = test_stage1_soln['class'].map(lambda x: int(x[-1]))

test_stage1_soln['ID'] = stage1_soln_ids

# test_s1_sol['class'] = test_s1_sol.idxmax(axis=1)
# test_s1_sol['class'] = test_s1_sol['class'].str[-1].astype(int)
# test_s1_sol['class'] = test_s1_sol['class']

test_stage1_soln.tail()

test_stage1_soln.head()

train_stage1.head()

train_stage1 = pd.merge(train_stage1, trainx_stage1, how='left', on='ID').fillna('')


test_stage1 = pd.merge(test_stage1, testx_stage1, how='left', on='ID').fillna('')
#pid = test['ID'].values

train_stage1.head()

test_stage1.head()

train_stage1.shape

test_stage1.shape

train_stage1.tail()

new_test = test_stage1[test_stage1.ID.isin(stage1_soln_ids)]

new_test.head()

new_test = new_test.drop('ID',axis=1)

new_test_ids = [i for i in range(3321, 3321+new_test.shape[0])]

len(new_test_ids)

new_test['ID'] = new_test_ids

test_stage1_soln.head()

new_test.shape

test_stage1_soln.shape

new_test['Class'] = test_stage1_soln['class'].values

new_test.shape

new_test.head()

new_test.tail()

train_stage1.head()

train_stage1.Class.value_counts()

train_stage1.shape

train_stage2 = pd.concat([train_stage1, new_test], axis=0)

train_stage2.head()

train_stage2.tail()

len(train_stage2.ID.unique())

train_stage2.shape

train_stage2.Class.value_counts()

len(train_stage2.ID.unique())

feather.write_dataframe(train_stage2, '../cache/train_stage2.feather')

train_stage2 = feather.read_dataframe('../cache/train_stage2.feather')

train_stage2.Class.value_counts()

y_class = train_stage2.Class.values

# from imblearn.over_sampling import RandomOverSampler

train_stage2.ID[train_stage2['Class'] == 8]

train_stage2[train_stage2.ID == 121]

# sampler = RandomOverSampler()
# X_res, y_res = sampler.fit_sample(ts_2, y_class)

train_stage2 = train_stage2.drop('ID', axis=1)

max_rows = 1054 # chosen because class 7 has 955 rows in stage2 train set

s = train_stage2.Class.value_counts()

num_classes = 9
rows_to_append = []
ids = {}
for i in tqdm(range(1, num_classes+1)):
    class_rows = max_rows - s[i]
    rows_to_append.append(class_rows)
    ids[i] = train_stage2.index[train_stage2['Class'] == i]

rows_to_append

import math

for i in tqdm(range(len(rows_to_append))):
    if rows_to_append[i] > 0:
        to_append = train_stage2.index.isin(ids[i+1])
#         print(to_append)
        n = to_append[to_append==True].sum()
        if n > rows_to_append[i]:
            m = rows_to_append[i]
            df = train_stage2.loc[to_append].copy()
            df = df[0:m-1]
            train_stage2 = pd.concat([train_stage2, df],axis=0)
        else:
            m = math.floor(rows_to_append[i]/n)
            for j in range(m):
                df = train_stage2.loc[to_append].copy()
                train_stage2 = pd.concat([train_stage2, df],axis=0)
            
        #df_to_append = train_stage2.loc[to_append].copy()
        #train_stage2 = pd.concat([train_stage2,df_to_append],axis=0)

train_stage2.Class.value_counts()

train_stage2.shape

train_stage2.head()

train_stage2.Class.value_counts()

feather.write_dataframe(train_stage2, '../cache/train_stage2_os.feather')

test_stage2 = pd.read_csv('../data/stage2_test_variants.csv')
testx_stage2 = pd.read_csv('../data/stage2_test_text.csv', sep="\|\|", engine='python', header=None, skiprows=1, names=["ID","Text"])

test_stage2.head()

testx_stage2.head()

test = pd.merge(test_stage2, testx_stage2, how='left', on='ID').fillna('')

test.head()

test.shape

len(set(test.Gene) -set(train_stage2.Gene))

len(set(test.Variation) -set(train_stage2.Variation))

feather.write_dataframe(test, '../cache/test_stage2.feather')
