==> notebooks/5_0_stage2_xgb.ipynb <==
import re
import numpy as np
import pandas as pd
import feather
import xgboost as xgb
import feather

from sklearn.model_selection import train_test_split
from sklearn.metrics import log_loss, confusion_matrix, f1_score, accuracy_score

from utils import print_confusion_matrix

def get_wt(cls, df):
    tot_neg_instances = df.shape[0] - p[cls]
    tot_pos_instances = p[cls]
    return float(tot_neg_instances/tot_pos_instances)

def get_wt2(cls, tot):
    tot_neg_instances = tot - p[cls]
    tot_pos_instances = p[cls]
    return float(tot_neg_instances/tot_pos_instances)

df_train = np.load('../cache/train_stage2_fe2.npy')

df_train.shape

df_test = np.load('../cache/test_stage2_fe2.npy')

df_test.shape

df = pd.read_csv('../cache/stage2_test_id.csv')

pid = df.ID

df1 = pd.read_csv('../cache/stage2_labels.csv')

df1.head()

y = df1['y'].values

y = y - 1 #fix for zero bound array

y

wts = np.load('../cache/stage2_train_weights.npy')

wts.shape

wts_per_class = np.load('../cache/stage2_train_weights_per_class.npy')
wts_per_class = wts_per_class.tolist()

print(wts_per_class)

wts_per_class[2]

 

denom = 0
fold = 10 
for i in range(fold):
    params = {
        'eta': 0.03333,
        'max_depth': 6,
        'subsample' : 0.8,
        'colsample_bytree':0.8,
        'objective': 'multi:softprob',
        'eval_metric': 'mlogloss',
        'num_class': 9,
        'seed': i,
        'tree_method': 'gpu_hist',
        'silent': True
    }
    x1, x2, y1, y2 = train_test_split(df_train, y, test_size=0.2, random_state=i, stratify=y, shuffle=True)
    
    w1 = [wts_per_class[j+1] for j in y1]
    w2 = [wts_per_class[j+1] for j in y2]
    watchlist = [(xgb.DMatrix(x1, y1, weight=w1), 'train'), (xgb.DMatrix(x2, y2, weight=w2), 'valid')]
    model = xgb.train(params, xgb.DMatrix(x1, y1, weight=w1), 1000,  watchlist, 
                      verbose_eval=50, early_stopping_rounds=100)
    pred_val =  model.predict(xgb.DMatrix(x2), ntree_limit=model.best_ntree_limit)
    score1 = log_loss(y2, pred_val, labels = list(range(9)))
    
    print('fold = {:d}'.format(i))
    print('val multi_log_loss: {}'.format(score1))
    
    fscore = f1_score(y2, pred_val.argmax(axis=1), labels = list(range(9)), average='macro')
    print('val f1_score: {}'.format(fscore))
    
    acc = accuracy_score(y2, pred_val.argmax(axis=1))
    print('val accuracy: {}'.format(acc))
    
    print(confusion_matrix(y2, pred_val.argmax(axis=1), labels = list(range(9))))
    
    print('-------------------')
    print('\n\n')
    #if score < 0.9:
    if denom != 0:
        pred = model.predict(xgb.DMatrix(df_test), ntree_limit=model.best_ntree_limit+80)
        preds += pred
    else:
        pred = model.predict(xgb.DMatrix(df_test), ntree_limit=model.best_ntree_limit+80)
        preds = pred.copy()
    denom += 1
    submission = pd.DataFrame(pred, columns=['class'+str(c+1) for c in range(9)])
    submission['ID'] = pid
    submission.to_csv('../submissions/sub5_0_stage2_xgb_fold_'  + str(i) + '.csv', index=False)


preds1 = preds/denom

preds1 = np.clip(preds1, a_min=0.05, a_max=0.95)

submission = pd.DataFrame(preds1, columns=['class'+str(c+1) for c in range(9)])
submission['ID'] = pid
submission.to_csv('../submissions/sub5_0_stage2_all_2_2_xgb.csv', index=False)
# scored 2.82570 on stage2 private LB, 1.70018 on stage2 public LB
# scored 2.81200 on stage2 private LB, 1.71324 on stage2 public LB (with new feature engg)
# scored 2.46932 on stage2 private LB, 1.33261 on stage2 public LB (without stratify and shuffle in train_test_split)
