==> notebooks/5_0_stage2_xgb_gridcv.ipynb <==
import sys
sys.path.insert(0, '/home/watts/Software/xgboost/python-package') # for xgboost

import re
import numpy as np
import pandas as pd
import feather
import xgboost as xgb
import feather

from sklearn.model_selection import train_test_split
from sklearn.metrics import log_loss

from sklearn.cross_validation import *
from sklearn.grid_search import GridSearchCV, RandomizedSearchCV

import pickle


def save_classifier(fname, clf):
    # save the classifier
    with open(fname, 'wb') as fid:
        pickle.dump(clf, fid)

def load_classifier(fname):
    # load it again
    with open(fname, 'rb') as fid:
        clf = pickle.load(fid)
        return clf

df_train = np.load('../cache/train_stage2_fe2.npy')

df1 = pd.read_csv('../cache/stage2_labels.csv')
y = df1['y'].values

df_test = np.load('../cache/test_stage2_fe2.npy')
df = pd.read_csv('../cache/stage2_test_id.csv')
pid = df.ID.values

df_train.shape

print(y)

y = y - 1 #fix for zero bound array

print(y.shape)

np.unique(y)

wts = np.load('../cache/stage2_train_weights.npy')

# wts_per_class = np.load('../cache/stage2_train_weights_per_class.npy')
# wts_per_class = wts_per_class.tolist()

xgb_model = xgb.XGBClassifier(objective='multi:softprob', sample_weight=wts)
xgbParams = xgb_model.get_xgb_params()
xgbParams['num_class'] = 9
parameters = {
    'learning_rate': [0.03, 0.035, 0.05], #so called `eta` value
    'max_depth': [5,6,10],
    'min_child_weight': [1,5,10],
    'silent': [1],
    'subsample': [0.8, 0.7],
    'colsample_bytree': [0.8, 0.7],
    'missing':[-999],
    'seed': [1337]
}


clf = GridSearchCV(xgb_model, parameters, n_jobs=5, 
                   cv=StratifiedKFold(y, 5),
                   scoring = 'neg_log_loss',
                   verbose=3, refit=True)


# neg_log_loos:-1.336173588372656
# {'colsample_bytree': 0.7, 'learning_rate': 0.05, 'max_depth': 5, 'min_child_weight': 1,
#  'missing': -999, 'seed': 1337, 'silent': 1, 'subsample': 0.8}
# 2.77967
# 1.71664

xgb_model = xgb.XGBClassifier(objective='multi:softprob', sample_weight=wts)
xgbParams = xgb_model.get_xgb_params()
xgbParams['num_class'] = 9
parameters = {
    'learning_rate': [0.03, 0.035, 0.05], #so called `eta` value
    'max_depth': [5,6,10],
    'min_child_weight': [1,5,10],
    'silent': [1],
    'subsample': [0.8, 0.7],
    'colsample_bytree': [0.8, 0.7],
    'missing':[-999],
    'seed': [1337]
}


clf = RandomizedSearchCV(xgb_model, parameters, n_jobs=5, 
                   cv=StratifiedKFold(y, 5, shuffle=True),
                   scoring = 'neg_log_loss',
                   verbose=3, refit=True)


clf.fit(df_train, y)

print(clf.best_score_)
print(clf.best_params_)

# f1_macro
# 0.4293248036851208
# {'colsample_bytree': 0.8, 'learning_rate': 0.05, 'max_depth': 10, 'min_child_weight': 1, 
#  'missing': -999, 'seed': 1337, 'silent': 1, 'subsample': 0.8}
# 2.74021, 1.72588

# neg_log_loos:-1.336173588372656
# {'colsample_bytree': 0.7, 'learning_rate': 0.05, 'max_depth': 5, 'min_child_weight': 1,
#  'missing': -999, 'seed': 1337, 'silent': 1, 'subsample': 0.8}
# 2.77967
# 1.71664

# -1.3876712019506041 (whien using wts, use xgboost from github)
# {'colsample_bytree': 0.7, 'learning_rate': 0.03, 'max_depth': 5, 
#  'min_child_weight': 5, 'missing': -999, 'seed': 1337, 'silent': 1, 'subsample': 0.8}
# private LB: 2.44337, pub LB: 1.69402

# -1.4902400080711418
# {'colsample_bytree': 0.7, 'learning_rate': 0.04, 'max_depth': 6, 'min_child_weight': 1, 
#  'missing': -999, 'seed': 1337, 'silent': 1, 'subsample': 0.7}
# private LB 2.56616
# public LB: 1.71541

# pip xgboost -1.4811547760008215
# {'colsample_bytree': 0.7, 'learning_rate': 0.03, 'max_depth': 5, 'min_child_weight': 5, 'missing': -999, 'seed': 1337, 'silent': 1, 'subsample': 0.8}
# In [14]:

# accuracy
# 0.5538086202222825
# {'colsample_bytree': 0.8, 'learning_rate': 0.05, 'max_depth': 10, 'min_child_weight': 1, 'missing': -999, 'seed': 1337, 'silent': 1, 'subsample': 0.8}
# In [26]:

# print(clf.best_score_)
# print(clf.best_params_)
#best score -1.42 (  grid with wts prob not used)

# print(clf.best_score_)
# print(clf.best_params_)
# best score: -1.40 (no wts)

best_parameters, score, _ = max(clf.grid_scores_, key=lambda x: x[1])
# print('Raw AUC score:', score)
for param_name in sorted(best_parameters.keys()):
    print("%s: %r" % (param_name, best_parameters[param_name]))



test_probs = clf.predict_proba(df_test)

test_probs.shape

print(test_probs)

test_probs = np.clip(test_probs, a_min=0.05, a_max=0.95)

submission = pd.DataFrame(test_probs, columns=['class'+str(c+1) for c in range(9)])
submission['ID'] = pid
# submission.to_csv('../submissions/subm_xgb_stage2_with_gridcv.csv', index=False)
# scored 1.70423 on the public leaderboard, 2.40243 on the private leaderboard


submission.to_csv('../submissions/subm_xgb_stage2_with_wt_gridcv.csv', index=False)
# scored 1.36438 on the public leaderboard, 2.35876 on the private leaderboard
