==> notebooks/5_1_stage2_RF.ipynb <==
import sys
sys.path.insert(0,'/home/watts/Software')

import re
import numpy as np
import pandas as pd
import feather
import xgboost as xgb
import feather

from sklearn.model_selection import train_test_split
from sklearn.metrics import log_loss, confusion_matrix, f1_score, accuracy_score

from sklearn.cross_validation import *
from sklearn.grid_search import GridSearchCV

from sklearn.ensemble import RandomForestClassifier
from sklearn.multioutput import MultiOutputClassifier

from sklearn.utils import shuffle

import pickle


def save_classifier(fname, clf):
    # save the classifier
    with open(fname, 'wb') as fid:
        pickle.dump(clf, fid)

def load_classifier(fname):
    # load it again
    with open(fname, 'rb') as fid:
        clf = pickle.load(fid)
        return clf

df_train = np.load('../cache/train_stage2_fe2.npy')
df1 = pd.read_csv('../cache/stage2_labels.csv')
y = df1['y'].values

df_test = np.load('../cache/test_stage2_fe2.npy')
df = pd.read_csv('../cache/stage2_test_id.csv')
pid = df.ID.values

len(pid)

df_train.shape

df_test.shape

# y = y - 1 #fix for zero bound array

print(y.shape)

np.unique(y)

wts_per_class = np.load('../cache/stage2_train_weights_per_class.npy')
wts_per_class = wts_per_class.tolist()
print(wts_per_class)

param_grid = { 
    'n_estimators': [50, 200, 700],
    'max_features': ['auto', 'sqrt', 'log2']
}

model_RF = RandomForestClassifier(random_state=0, class_weight='balanced')
clf = GridSearchCV(model_RF, param_grid=param_grid, scoring='neg_log_loss',cv=5,n_jobs=4)
clf.fit(df_train, y)
fname = '../cache/rf_multi_clf.pkl'
save_classifier(fname, clf)

print(clf.best_score_)
print(clf.best_params_)

# with my calculated weights
# -1.5031283568737968
# {'max_features': 'log2', 'n_estimators': 700}

# with balanced weights
# -1.4858419408875922
# {'max_features': 'auto', 'n_estimators': 700}

pred_proba = clf.predict_proba(df_test)

pred_proba = pred_proba.clip(min=0.05, max=0.95)

print(pred_proba.shape)

n_classes = np.unique(y)

df = pd.DataFrame(pred_proba)
l = []
for cls in range(len(n_classes)):
    l.append('class'+str(cls+1))
df.columns = l

df.shape

df['ID'] = pid

df.head()

# df.to_csv('../submissions/sub_stage2_mop_rf.csv', index=False)
# scored 2.60615 on stage2 private LB and 1.61225 on stage2 public LB with weights


df.to_csv('../submissions/sub_stage2_wt_rf.csv', index=False)
# 2.20910 on stage2 private LB, 1.58565 on stag2 public LB

df1 = pd.DataFrame()
df1['ID'] = pid
df1.to_csv('../cache/test_id.csv', index=False)

df = pd.DataFrame()
df['y'] = y
df.to_csv('../cache/train_stage2_labels.csv', index=False)

X = df_train
Y = y
Z = df_test


df.shape

x1 = np.load('../cache/train_stage2_x1.npy')
x2 = np.load('../cache/train_stage2_x2.npy')
y1 = np.load('../cache/train_stage2_y1.npy')
y2 = np.load('../cache/train_stage2_y2.npy')

pd.value_counts(y2)

y2.shape

(2 * (221/738) *0.11)/((221/738) + 0.11) 

w1 = np.array([wts_per_class[j] for j in y1], )
w2 = np.array([wts_per_class[j] for j in y2], )

clf1 = RandomForestClassifier(random_state=0, max_features='auto', n_estimators=700, class_weight='balanced')



clf1.fit(x1, y1)


test_preds = clf1.predict_proba(x2)
print(test_preds)

score2 = log_loss(y2, test_preds, labels = list(range(1,10)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(y2, test_preds.argmax(axis=1)+1, labels = list(range(1,10)), average='micro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(y2, test_preds.argmax(axis=1)+1)
print('final accuracy: {}'.format(acc))

print(confusion_matrix(y2, test_preds.argmax(axis=1)+1, labels = list(range(1,10))))

# with my weights
# final multi_log_loss: 1.0051730722455978
# final f1_score: 0.7073170731707317
# final accuracy: 0.7073170731707317
# [[ 97   0   1  19   5   1   6   0   0]
#  [  4  43   0   1   0   2  56   0   0]
#  [  0   0   4   6   1   0  13   0   0]
#  [ 12   1   1 114   0   0  13   0   0]
#  [ 16   1   0   4  26   2   5   0   0]
#  [  8   0   0   2   2  34   9   0   0]
#  [  1  13   2   3   2   0 200   0   0]
#  [  0   0   0   0   0   0   2   0   0]
#  [  0   0   0   0   0   0   2   0   4]]

# with 'balanced' weights
# final multi_log_loss: 0.9154469754860213
# final f1_score: 0.6964769647696477
# final accuracy: 0.6964769647696477
# [[ 86   1   2  25  14   3   2   0   0]
#  [  2  51   0   2   0   0  44   1   0]
#  [  1   3   8   5   0   0   2   0   0]
#  [ 17   0   1 119   1   1  11   0   0]
#  [ 10   3   1   7  21   3   8   0   0]
#  [  6   0   1   5   1  37   9   0   0]
#  [  2   8   7   3   5   0 185   0   1]
#  [  1   0   0   0   0   0   3   0   0]
#  [  0   0   0   1   0   0   1   0   7]]

import numpy as np
import matplotlib.pyplot as plt

# conf_arr = [[33,2,0,0,0,0,0,0,0,1,3], 
#             [3,31,0,0,0,0,0,0,0,0,0], 
#             [0,4,41,0,0,0,0,0,0,0,1], 
#             [0,1,0,30,0,6,0,0,0,0,1], 
#             [0,0,0,0,38,10,0,0,0,0,0], 
#             [0,0,0,3,1,39,0,0,0,0,4], 
#             [0,2,2,0,4,1,31,0,0,0,2],
#             [0,1,0,0,0,0,0,36,0,2,0], 
#             [0,0,0,0,0,0,1,5,37,5,1], 
#             [3,0,0,0,0,0,0,0,0,39,0], 
#             [0,0,0,0,0,0,0,0,0,0,38]]

conf_arr = [[97,0,1,19,5,1,6,0,0],
            [4,43,0,1,0,2,56,0,0],
            [0,0,4,6,1,0,13,0,0],
            [12,1,1,114,0,0,13,0,0],
            [16,1,0,4,26,2,5,0,0],
            [8,0,0,2,2,34,9,0,0],
            [1,13,2,3,2,0,200,0,0],
            [0,0,0,0,0,0,2,0,0],
            [0,0,0,0,0,0,2,0,4]]



conf_arr.shape

norm_conf = []
for i in conf_arr:
    a = 0
    tmp_arr = []
    a = sum(i, 0)
    for j in i:
        tmp_arr.append(float(j)/float(a))
    norm_conf.append(tmp_arr)

fig = plt.figure()
plt.clf()
ax = fig.add_subplot(111)
ax.set_aspect(1)
res = ax.imshow(np.array(norm_conf), cmap=plt.cm.jet, 
                interpolation='nearest')

width, height =[9,9]

for x in range(width):
    for y in range(height):
        ax.annotate(str(conf_arr[x][y]), xy=(y, x), 
                    horizontalalignment='center',
                    verticalalignment='center')

cb = fig.colorbar(res)
alphabet = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
plt.xticks(range(width), alphabet[:width])
plt.yticks(range(height), alphabet[:height])
plt.savefig('../cache/confusion_matrix_rf.png', format='png')
