==> notebooks/5_1_stage2_xgb_gridcv.ipynb <==
# import sys
# sys.path.insert(0,'/home/watts/Software')

import re
import numpy as np
import pandas as pd
import feather
import xgboost as xgb
import feather

from sklearn.cross_validation import StratifiedKFold
# from sklearn.cross_validation import *
from sklearn.grid_search import GridSearchCV

from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer

import pickle

from sklearn.metrics import log_loss, confusion_matrix, f1_score, accuracy_score

# Any results you write to the current directory are saved as output.
submission = pd.read_csv('../data/stage2_sample_submission.csv')
stage1_test = pd.read_csv('../data/test_variants')
stage2_test = pd.read_csv('../data/stage2_test_variants.csv')
stage1_solution = pd.read_csv('../data/stage1_solution_filtered.csv')

stage1_solution = stage1_solution.merge(stage1_test, how = 'left', on = 'ID')



wts_per_class = np.load('../cache/stage2_train_weights_per_class.npy')
wts_per_class = wts_per_class.tolist()
print(wts_per_class)

df_train = np.load('../cache/train_stage2_fe2.npy')
df1 = pd.read_csv('../cache/stage2_labels.csv')
y = df1['y'].values

df_test = np.load('../cache/test_stage2_fe2.npy')
df = pd.read_csv('../cache/stage2_test_id.csv')
pid = df.ID.values

y = y - 1 #fix for zero bound array

print(y.shape)

np.unique(y)

w = np.array([wts_per_class[j+1] for j in y], )

xgb_model = xgb.XGBClassifier(objective='multi:softprob', sample_weight=w)
xgbParams = xgb_model.get_xgb_params()
xgbParams['num_class'] = 9
parameters = {
              'learning_rate': [0.02, 0.01, 0.03], #so called `eta` value
              'max_depth': [5,6],
              'min_child_weight': [5,11],
              'silent': [1],
              'subsample': [0.8, 0.8],
              'colsample_bytree': [0.7, 0.8],
              'missing':[-999],
              'seed': [1337]}


clf = GridSearchCV(xgb_model, parameters, n_jobs=10, 
                   cv=StratifiedKFold(y, 5, True),
                   scoring = 'neg_log_loss',
                   verbose=2, refit=True)

clf.fit(df_train, y) 

print(clf.best_score_)
print(clf.best_params_)

# -1.4032872189073649
# {'colsample_bytree': 0.7, 'learning_rate': 0.03, 'max_depth': 6, 'min_child_weight': 11,
#  'missing': -999, 'nthread': 4, 'seed': 1337, 'silent': 1, 'subsample': 0.8}
# scored 1.70423 on the public leaderboard, 2.40243 on the private leaderboard

# -0.9878011259178897
# {'colsample_bytree': 0.8, 'learning_rate': 0.03, 'max_depth': 6, 'min_child_weight': 5, 
#  'missing': -999, 'seed': 1337, 'silent': 1, 'subsample': 0.8}
# 2.39785 stage2 private, 1.76875 stage2 public

best_parameters, score, _ = max(clf.grid_scores_, key=lambda x: x[1])
for param_name in sorted(best_parameters.keys()):
    print("%s: %r" % (param_name, best_parameters[param_name]))



test_probs = clf.predict_proba(df_test)

test_probs.shape

print(test_probs)

test_probs = test_probs.clip(min=0.05, max=0.95) # hack for kaggle since kaggle punishes severly if you predict 
# 1 and it is 0

submission = pd.DataFrame(test_probs, columns=['class'+str(c+1) for c in range(9)])
submission['ID'] = pid
submission.to_csv('../submissions/subm_xgb_stage2_with_gridcv.csv', index=False)


# scored 1.70423 on the public leaderboard, 2.40243 on the private leaderboard

save_classifier('../cache/xgb_clf.pkl', clf)

# now get the CV score

x1 = np.load('../cache/train_stage2_x1.npy')
x2 = np.load('../cache/train_stage2_x2.npy')
y1 = np.load('../cache/train_stage2_y1.npy')
y2 = np.load('../cache/train_stage2_y2.npy')

w1 = np.array([wts_per_class[j] for j in y1], )
w2 = np.array([wts_per_class[j] for j in y2], )

params = {
    'objective': 'multi:softprob', 'eval_metric': 'mlogloss',
    'colsample_bytree': 0.8, 'learning_rate': 0.03, 'max_depth': 6, 
    'min_child_weight': 5, 'missing': -999, 'nthread': 4, 'seed': 1337, 
    'silent': 1, 'subsample': 0.8
}
clf = xgb.XGBClassifier(**params)

np.unique(y2)

y1= y1 - 1 #fix for zero bound array
y2 = y2-1

clf.fit(x1, y1, sample_weight=w1)

test_preds = clf.predict_proba(x2)
print(test_preds)

score2 = log_loss(y2, test_preds, labels = list(range(9)))
print('final multi_log_loss: {}'.format(score2))

fscore = f1_score(y2, test_preds.argmax(axis=1), labels = list(range(9)), average='micro')
print('final f1_score: {}'.format(fscore))

acc = accuracy_score(y2, test_preds.argmax(axis=1))
print('final accuracy: {}'.format(acc))

print(confusion_matrix(y2, test_preds.argmax(axis=1), labels = list(range(9))))
