==> notebooks/5_2_stage2_multiclassifers_org.ipynb <==
import numpy as np
import pandas as pd
import feather

from sklearn.multiclass import OneVsRestClassifier
from sklearn.multiclass import OneVsOneClassifier
from sklearn.multiclass import OutputCodeClassifier
  
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
  
from sklearn.svm import LinearSVC
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import (LinearRegression, Lasso, ElasticNet, Ridge,
                                  Perceptron)
from sklearn.tree import DecisionTreeClassifier
from sklearn.grid_search import GridSearchCV
from sklearn import svm
from sklearn.linear_model import Perceptron
from sklearn.naive_bayes import MultinomialNB

from sklearn.utils.testing import assert_array_equal
from sklearn.utils.testing import assert_equal
from sklearn.utils.testing import assert_almost_equal
from sklearn.utils.testing import assert_true
from sklearn.utils.testing import assert_false
from sklearn.utils.testing import assert_raises
  
from sklearn.utils.testing import assert_greater
import pickle

from sklearn.metrics import log_loss, confusion_matrix, f1_score, accuracy_score

df_train = np.load('../cache/train_stage2_fe2.npy')

df_test = np.load('../cache/test_stage2_fe2.npy')

df = pd.read_csv('../cache/train_stage2_labels.csv')

df.head()

y = df['y'].values

x1 = np.load('../cache/train_stage2_x1.npy')
x2 = np.load('../cache/train_stage2_x2.npy')
y1 = np.load('../cache/train_stage2_y1.npy')
y2 = np.load('../cache/train_stage2_y2.npy')

wts_per_class = np.load('../cache/stage2_train_weights_per_class.npy')
wts_per_class = wts_per_class.tolist()
print(wts_per_class)

w1 = np.array([wts_per_class[j] for j in y1], )
w2 = np.array([wts_per_class[j] for j in y2], )

w = np.array([wts_per_class[j] for j in y], )

ovo = OneVsOneClassifier(LinearSVC(random_state=0))
Cs = [0.1, 0.01, 0.03, 0.05, 0.08, 0.5, 0.8]
cv = GridSearchCV(ovo, {'estimator__C': Cs}, verbose=2, refit=True)
cv.fit(df_train, y)
best_C = cv.best_estimator_.estimators_[0].C
assert_true(best_C in Cs)


print(best_C)

test_preds = cv.predict(df_test)
print(test_preds)

df = pd.DataFrame()
df['Class1'] =0
df['Class2'] =0
df['Class3'] =0
df['Class4'] =0
df['Class5'] =0
df['Class6'] =0
df['Class7'] =0
df['Class8'] =0
df['Class9'] =0

df['Class1'] = [0.95 if i == 0 else 0.05 for i in test_preds]
df['Class2'] = [0.95 if i == 1 else 0.05 for i in test_preds]
df['Class3'] = [0.95 if i == 2 else 0.05 for i in test_preds]
df['Class4'] = [0.95 if i == 3 else 0.05 for i in test_preds]
df['Class5'] = [0.95 if i == 4 else 0.05 for i in test_preds]
df['Class6'] = [0.95 if i == 5 else 0.05 for i in test_preds]
df['Class7'] = [0.95 if i == 6 else 0.05 for i in test_preds]
df['Class8'] = [0.95 if i == 7 else 0.05 for i in test_preds]
df['Class9'] = [0.95 if i == 8 else 0.05 for i in test_preds]

df.head()

df1 = pd.read_csv('../cache/test_id.csv')

df1.head()

df['ID'] = df1['ID'].values

df.to_csv('../submissions/sub_svc_one_on_one.csv', index=None)
# 2.25939 on stage2 private LB, 2.90524 on stage2 public LB

clf = OneVsOneClassifier(LinearSVC(random_state=0, C=0.1))

clf.fit(x1,y1)

test_preds = clf.predict(x2)
print(test_preds)

y2

from utils import *
score2 = log_loss(y2, test_preds)
# score2 = multi_log_loss(y2, test_preds, labels = list(range(1,10)))
print('CV multi_log_loss: {}'.format(score2))

fscore = f1_score(y2, test_preds, average='macro')
print('CV f1_score: {}'.format(fscore))

acc = accuracy_score(y2, test_preds)
print('CV accuracy: {}'.format(acc))

print(confusion_matrix(y2, test_preds, labels = list(range(1,10))))

clf = OneVsRestClassifier(svm.SVC()).fit(df_train, y)
Y_pred = clf.predict(df_test)


save_classifier('../cache/svm_svr_clf.pkl', clf)

print(Y_pred)

df = pd.DataFrame()
df['Class1'] =0
df['Class2'] =0
df['Class3'] =0
df['Class4'] =0
df['Class5'] =0
df['Class6'] =0
df['Class7'] =0
df['Class8'] =0
df['Class9'] =0

df['Class1'] = [0.95 if i == 0 else 0.05 for i in Y_pred]
df['Class2'] = [0.95 if i == 1 else 0.05 for i in Y_pred]
df['Class3'] = [0.95 if i == 2 else 0.05 for i in Y_pred]
df['Class4'] = [0.95 if i == 3 else 0.05 for i in Y_pred]
df['Class5'] = [0.95 if i == 4 else 0.05 for i in Y_pred]
df['Class6'] = [0.95 if i == 5 else 0.05 for i in Y_pred]
df['Class7'] = [0.95 if i == 6 else 0.05 for i in Y_pred]
df['Class8'] = [0.95 if i == 7 else 0.05 for i in Y_pred]
df['Class9'] = [0.95 if i == 8 else 0.05 for i in Y_pred]

df['ID'] = df1['ID'].values

df.to_csv('../submissions/sub_svm_one_vs_rest.csv', index=None)

# 2.54206 on stage2 private leaderboard, 3.16635 on stage2 public leaderboard

clf.fit(x1,y1)
test_preds = clf.predict(x2)

score2 = log_loss(y2, test_preds)
# score2 = multi_log_loss(y2, test_preds, labels = list(range(1,10)))
print('CV multi_log_loss: {}'.format(score2))

fscore = f1_score(y2, test_preds, average='macro')
print('CV f1_score: {}'.format(fscore))

acc = accuracy_score(y2, test_preds)
print('CV accuracy: {}'.format(acc))

print(confusion_matrix(y2, test_preds, labels = list(range(1,10))))

ovo = OneVsOneClassifier(MultinomialNB())
Y_pred =ovo.fit(abs(df_train), y).predict(df_test)

df = pd.DataFrame()
df['Class1'] =0
df['Class2'] =0
df['Class3'] =0
df['Class4'] =0
df['Class5'] =0
df['Class6'] =0
df['Class7'] =0
df['Class8'] =0
df['Class9'] =0

df['Class1'] = [0.95 if i == 0 else 0.05 for i in Y_pred]
df['Class2'] = [0.95 if i == 1 else 0.05 for i in Y_pred]
df['Class3'] = [0.95 if i == 2 else 0.05 for i in Y_pred]
df['Class4'] = [0.95 if i == 3 else 0.05 for i in Y_pred]
df['Class5'] = [0.95 if i == 4 else 0.05 for i in Y_pred]
df['Class6'] = [0.95 if i == 5 else 0.05 for i in Y_pred]
df['Class7'] = [0.95 if i == 6 else 0.05 for i in Y_pred]
df['Class8'] = [0.95 if i == 7 else 0.05 for i in Y_pred]
df['Class9'] = [0.95 if i == 8 else 0.05 for i in Y_pred]
df['ID'] = df1['ID'].values
df.to_csv('../submissions/sub_nb_one_on_one.csv', index=None)

# 2.89574 on stage2 private leaderboard, 2.95424 on stage2 public leaderboard

ovo.fit(abs(x1),y1)
test_preds = ovo.predict(x2)

score2 = log_loss(y2, test_preds)
# score2 = multi_log_loss(y2, test_preds, labels = list(range(1,10)))
print('CV multi_log_loss: {}'.format(score2))

fscore = f1_score(y2, test_preds, average='weighted')
print('CV f1_score: {}'.format(fscore))

acc = accuracy_score(y2, test_preds)
print('CV accuracy: {}'.format(acc))

print(confusion_matrix(y2, test_preds, labels = list(range(1,10))))
