==> notebooks/5_2_stage2_RF.ipynb <==
import re
import numpy as np
import pandas as pd
import feather
import xgboost as xgb
import feather
from sklearn.base import BaseEstimator as be
from sklearn.base import TransformerMixin as tm
from sklearn.pipeline import Pipeline
from sklearn.pipeline import FeatureUnion
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import HashingVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.decomposition import TruncatedSVD
from sklearn.model_selection import train_test_split
from sklearn.metrics import log_loss

from sklearn.cross_validation import *
from sklearn.grid_search import GridSearchCV

from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer

from sklearn.ensemble import RandomForestClassifier
from sklearn.multioutput import MultiOutputClassifier

from sklearn.utils import shuffle

import pickle


def save_classifier(fname, clf):
    # save the classifier
    with open(fname, 'wb') as fid:
        pickle.dump(clf, fid)

def load_classifier(fname):
    # load it again
    with open(fname, 'rb') as fid:
        clf = pickle.load(fid)
        return clf

# Any results you write to the current directory are saved as output.
submission = pd.read_csv('../data/stage2_sample_submission.csv')
stage1_test = pd.read_csv('../data/test_variants')
stage2_test = pd.read_csv('../data/stage2_test_variants.csv')
stage1_solution = pd.read_csv('../data/stage1_solution_filtered.csv')

stage1_solution = stage1_solution.merge(stage1_test, how = 'left', on = 'ID')



df_train = feather.read_dataframe('../cache/train_stage2_fe.feather')

train_ids = df_train.ID.values

df_train = df_train.drop('ID', axis=1)

y = df_train['Class'].values

df_test = feather.read_dataframe('../cache/test_stage2_fe.feather')
pid = df_test.ID
df_test = df_test.drop('ID', axis=1)
# df_test['Class'] = y_hack

df_train.shape

df_train.head()

df_test = feather.read_dataframe('../cache/test_stage2_fe.feather')

df_test.shape

df_test.head()

########################################
## process texts in datasets
########################################


# The function "text_to_wordlist" is from
# https://www.kaggle.com/currie32/quora-question-pairs/the-importance-of-cleaning-text
def text_to_wordlist(text, remove_stopwords=True, stem_words=True):
    # Clean the text, with the option to remove stopwords and to stem words.
    
    # Convert words to lower case and split them
    text = text.lower().split()

    my_stopwords = [
        "fig", "figure", "et", "al", "table",
        "data", "analysis", "analyze", "study",
        "method", "result", "conclusion", "author",
        "find", "found", "show", "perform",
        "demonstrate", "evaluate", "discuss"
    ]
    
    # Optionally, remove stop words
    if remove_stopwords:
        stops = set(stopwords.words("english"))
        text = [w for w in text if not w in stops]
        text = [w for w in text if not w in my_stopwords]
    
    text = " ".join(text)

    # Clean the text
    text = re.sub(r"[^A-Za-z0-9^,!.\/'+-=]", " ", text)
    text = re.sub(r"what's", "what is ", text)
    text = re.sub(r"\'s", " ", text)
    text = re.sub(r"\'ve", " have ", text)
    text = re.sub(r"can't", "cannot ", text)
    text = re.sub(r"n't", " not ", text)
    text = re.sub(r"i'm", "i am ", text)
    text = re.sub(r"\'re", " are ", text)
    text = re.sub(r"\'d", " would ", text)
    text = re.sub(r"\'ll", " will ", text)
    text = re.sub(r",", " ", text)
    text = re.sub(r"\.", " ", text)
    text = re.sub(r"!", " ! ", text)
    text = re.sub(r"\/", " ", text)
    text = re.sub(r"\^", " ^ ", text)
    text = re.sub(r"\+", " + ", text)
    text = re.sub(r"\-", " - ", text)
    text = re.sub(r"\=", " = ", text)
    text = re.sub(r"'", " ", text)
    text = re.sub(r"(\d+)(k)", r"\g<1>000", text)
    text = re.sub(r":", " : ", text)
    text = re.sub(r" e g ", " eg ", text)
    text = re.sub(r" b g ", " bg ", text)
    text = re.sub(r" u s ", " american ", text)
    text = re.sub(r"\0s", "0", text)
    text = re.sub(r" 9 11 ", "911", text)
    text = re.sub(r"e - mail", "email", text)
    text = re.sub(r"j k", "jk", text)
    text = re.sub(r"\s{2,}", " ", text)
    
    # Optionally, shorten words to their stems
    if stem_words:
        text = text.split()
        stemmer = SnowballStemmer('english')
        stemmed_words = [stemmer.stem(word) for word in text]
        text = " ".join(stemmed_words)
    
    # Return a list of words
    return(text)

# https://www.kaggle.com/the1owl/redefining-treatment-0-57456
class cust_regression_vals(be, tm):
    def fit(self, x, y=None):
        return self
    def transform(self, x):
        x = x.drop(['Gene', 'Variation', 'ID','Text', 'GeneVar'],axis=1).values
        return x

class cust_txt_col(be, tm):
    def __init__(self, key):
        self.key = key
    def fit(self, x, y=None):
        return self
    def transform(self, x):
        return x[self.key].apply(str)


# from https://www.kaggle.com/the1owl/redefining-treatment-0-57456
fp = Pipeline([
    ('union', FeatureUnion(
        n_jobs = -1,
        transformer_list = [
            ('standard', cust_regression_vals()),
            ('pi1', Pipeline([('Gene', cust_txt_col('Gene')), 
                                       ('count_Gene', CountVectorizer(analyzer=u'char',ngram_range=(1, 3))), 
                                       ('tsvd1', TruncatedSVD(n_components=20, n_iter=25, random_state=12))])),
            ('pi2', Pipeline([('Variation', cust_txt_col('Variation')), 
                                       ('count_Variation', CountVectorizer(analyzer=u'char',ngram_range=(1, 3))), 
                                       ('tsvd2', TruncatedSVD(n_components=20, n_iter=25, random_state=12))])),
            ('pi3', Pipeline([('GeneVar', cust_txt_col('GeneVar')), 
                                       ('count_GeneVar', CountVectorizer(analyzer=u'char', ngram_range=(1, 3))), 
                                       ('tsvd2', TruncatedSVD(n_components=20, n_iter=25, 
                                                                            random_state=12))])),
            ('pi4',Pipeline([('Text', cust_txt_col('Text')), 
                                       ('hv', HashingVectorizer(decode_error='ignore', 
                                                                                        n_features=2 ** 16,
                                                                                        non_negative=True, 
                                                                                        ngram_range=(1, 5))),
                                       ('tfidf_Text', TfidfTransformer()), 
                                       ('tsvd3', TruncatedSVD(n_components=300, n_iter=25, 
                                                                            random_state=12))]))

        
        ])
    )])


df_train['Text'] = [text_to_wordlist(w) for w in df_train['Text'].values]

df_test['Text'] = [text_to_wordlist(w) for w in df_test['Text'].values]

df_train.head()

y = df_train['Class'].values

df_train['ID'] = [i for i in range(df_train.shape[0])]

tr = df_train
te = df_test

# y = tr['Class'].values

set(te.columns) - set(tr.columns)

df_train = df_train.drop('Class',axis=1)

# df_train = df_train.drop('ID', axis=1)
# df_test = df_test.drop('ID', axis=1)

# df_train = tr

df_train.head()

df_train.shape

df_train = fp.fit_transform(df_train)
print (df_train.shape)


df_test = fp.fit_transform(df_test)
print (df_test.shape)

np.save('../cache/train_stage2_fe2', df_train)

np.save('../cache/test_stage2_fe2', df_test)

y = y - 1 #fix for zero bound array

print(y.shape)

np.unique(y)

# parameters = {
#   "n_estimators": [5, 50, 200],
#   "max_depth" : [None, 10,20],
#   "min_samples_split" : [2, 5, 10],
# }

param_grid = { 
    'n_estimators': [50, 200, 700],
    'max_features': ['auto', 'sqrt', 'log2']
}

for cls in [0,1,2,3,4,5,6,7,8]:
    y_this_class = [1 if cls==i else 0 for i in y]
#     print(y_this_class)
    model_RF = RandomForestClassifier(random_state=0)
    clf = GridSearchCV(model_RF, param_grid=param_grid, cv=5,n_jobs=4)
    clf.fit(df_train, y_this_class )
    fname = '../cache/rf_cls_' + str(cls) + '.pkl'
    save_classifier(fname, clf)

n_classes = np.unique(y)

df = pd.DataFrame()
for cls in n_classes:
    fname = '../cache/rf_cls_' + str(cls) + '.pkl'
    clf = load_classifier(fname)
    print(clf.best_score_)
    print(clf.best_params_)
    test_probs = clf.predict_proba(df_test)[:,1]
    df['class' + str(cls+1)] = test_probs

df['ID'] = pid

df.head()

df.to_csv('../submissions/submission_staeg2_rf.csv', index=False)


# scored 2.18400 on stage2 public leaderboard, and 8.10579 on private leaderboard


df1 = pd.DataFrame()
df1['ID'] = pid
df1.to_csv('../cache/test_id.csv', index=False)

df = pd.DataFrame()
df['y'] = y
df.to_csv('../cache/train_stage2_labels.csv', index=False)

X = df_train
Y = y
Z = df_test


Y1 = Y.T

y2 = shuffle(Y, random_state=1)
y3 = shuffle(Y, random_state=2)
y4 = shuffle(Y, random_state=3)
y5 = shuffle(Y, random_state=4)
y6 = shuffle(Y, random_state=5)
y7 = shuffle(Y, random_state=6)
y8 = shuffle(Y, random_state=7)
y9 = shuffle(Y, random_state=8)

Y1 = np.vstack((Y, y2, y3, y4, y5, y6, y7, y8, y9)).T

forest = RandomForestClassifier(n_estimators=1000, random_state=1)
multi_target_forest = MultiOutputClassifier(forest, n_jobs=-1)


preds = multi_target_forest.fit(X, Y1).predict(Z)

preds.shape

print(preds)

preds_proba = multi_target_forest.fit(X, Y1).predict_proba(Z)

preds_proba[0].shape

preds_proba[0]

preds_proba[0]

df = pd.DataFrame(preds_proba[0], 
                  columns=['class1', 'class2', 'class3', 'class4', 'class5', 'class6', 'class7', 'class8', 'class9'])


df['ID'] = pid
df.to_csv('../submissions/sub_stage2_multi_rf.csv', index=False)

# scored 2.64206 on stage2 private LB, 1.54953 on stage2 public LB

save_classifier('../cache/stage2_multi_rf.pkl', multi_target_forest)

df.head()

df.shape
